[kwinrules] Update from window menu when the KCM is already open
Add a DBus endpoint to parseArguments() method, so that
the window UUID passed by `Edit .. specific properties` menu
can be processed when the KCM is already open.

This allows for subsequent calls from different windows into
the same KCM instance. If the same window is called twice,
the corresponding rule will be detected by the matching mechanism.

BUG: 417923
FIXED-IN: 5.23