wayland: Maximize window if it doesn't fit the placement area
If the window is too large for the placement area, maximize it. Note
that there is a proposal upstream to extend the xdgs-shell protocol with
an event that can be used by the compositor to tell the client how big
the window can be. But it's worth to have this change in meanwhile,
until the relevant protocol change lands and all clients are ported to
it.

https://gitlab.freedesktop.org/wayland/wayland-protocols/-/merge_requests/41