Revert "effects: Link builtin effects with executables"
This reverts commit 141947d2e85f825a1c4d4300c0b8f012d08d3f1c.

As it turns out, it slows down Qt plugin loading quite a lot.

Initially, metadata didn't include translated strings but as soon as
translators started adding strings, kwin started freezing any time it
needed to load a QImage.

The root cause of the freeze is that QStaticPlugin::metaData() doesn't
cache json metadata. It performs cbor -> json conversion every time it
is called. That combined with a lot of translated strings results in
noticeable kwin slowdowns.